refactor hooks to common util
